backend/requirements.txt
fastapi
uvicorn
httpx
pydantic
PyJWT
python-dotenv
trafilatura
aiolimiter
python-dateutil
tiktoken
openai
orjson
transformers
#datasets
huggingface_hub
tokenizers
tensorflow
streamlit-aggrid
pyarrow==14.0.2
streamlit==1.28.2
python-jose[cryptography]==3.3.0
python-multipart==0.0.6
google-auth==2.23.4
google-auth-oauthlib==1.1.0


